Replace browserify in scripts:bundle task
