+++ bokehjs/gulp/linker.ts
@@ -6,61 +6,11 @@ import * as escodegen from "escodegen"
 import * as estraverse from "estraverse"
 import {Program, Node, CallExpression} from "estree"
 
+import {prelude, plugin_prelude} from "./prelude"
+
 const str = JSON.stringify
 const exists = fs.existsSync
 const is_dir = (path: string) => fs.statSync(path).isDirectory()
-
-const prelude = (entry: number, modules: string) => {
-  return `\
-(function(modules, entry) {
-  var cache = {};
-
-  var require = function(name) {
-    if (!cache[name]) {
-      if (!modules[name]) {
-        var err = new Error("Cannot find module '" + name + "'");
-        err.code = 'MODULE_NOT_FOUND';
-        throw err;
-      }
-
-      var module = cache[name] = {exports: {}};
-      modules[name].call(module.exports, require, module, module.exports);
-    }
-
-    return cache[name].exports;
-  }
-
-  var main = require(entry);
-  main.require = require;
-
-  main.register_plugin = function(plugin_modules, plugin_entry) {
-    for (var name in plugin_modules) {
-      modules[name] = plugin_modules[name];
-    }
-
-    var plugin = require(plugin_entry);
-
-    for (var name in plugin) {
-      main[name] = plugin[name];
-    }
-  }
-
-  return main;
-})(${modules}, ${entry});
-`
-}
-
-const plugin_prelude = (entry: number, modules: string) => {
-  return `\
-(function(modules, entry) {
-  if (Bokeh != null) {
-    Bokeh.register_plugin(modules, entry);
-  } else {
-    throw new Error("Cannot find Bokeh. You have to load it prior to loading plugins.");
-  }
-})(${modules}, ${entry});
-`
-}
 
 export interface LinkerOpts {
   entries: string[]
